backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel>=0.0.16
psycopg2-binary
python-dotenv==1.0.0
python-multipart==0.0.9
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

